refactor: Add public `MockApiKeyRepo` struct for tests
