+++ api/src/infrastructure/repositories/mod.rs
@@ -3,6 +3,9 @@ pub mod api_keys;
 pub mod monitor;
 pub mod monitor_repo;
 
+#[cfg(test)]
+pub mod mock_api_key_repo;
+
 use std::marker::{Send, Sync};
 
 use async_trait::async_trait;
